feat :sparkles: calculate rate installments for loans and depurate comments

+++ backLolaCredits/Models/Loan.cs
@@ -9,7 +9,7 @@ public class Loan
     public DateOnly LoanDate { get; set; }
     public DateOnly DueDate { get; set; }              // final payment date
     public int Months { get; set; }              // total number of installments
-    public int InterestRate { get; set; }        // percentage
+    public decimal InterestRate { get; set; }    // percentage (e.g., 4.5 for 4.5%)
     public int PaymentDay { get; set; }          // day of month for payments (1-28)
 
     public decimal MonthlyAmount { get; set; }  // calculated
